@@ -2,7 +2,7 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": "^8.1",
+        "php": "^8.3",
         "ext-iconv": "*",
         "ext-json": "*",
         "ext-simplexml": "*",
@@ -14,10 +14,11 @@
         "doctrine/orm": "^2.10",
         "flagception/flagception-bundle": "^4.1",
         "fresh/doctrine-enum-bundle": "^9.0",
-        "friendsofsymfony/elastica-bundle": "^6.0",
         "friendsofsymfony/jsrouting-bundle": "^3.0",
         "geocoder-php/nominatim-provider": "^5.2",
-        "jms/serializer-bundle": "^4.0",
+        "http-interop/http-factory-guzzle": "^1.2",
+        "jms/serializer-bundle": "^5.0",
+        "jsor/doctrine-postgis": "^2.1",
         "league/geotools": "^1.0",
         "maltehuebner/impress-bundle": "^0.2",
         "mhujer/breadcrumbs-bundle": "^1.5",
@@ -42,6 +43,7 @@
         "symfony/proxy-manager-bridge": "^6.2",
         "symfony/runtime": "^6.2",
         "symfony/serializer": "^6.2",
+        "symfony/string": "^6.2",
         "symfony/twig-bridge": "^6.2",
         "symfony/webpack-encore-bundle": "^1.7",
         "symfony/yaml": "^6.2",
@@ -50,7 +52,7 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^9.0",
-        "rector/rector": "^0.14.5",
+        "rector/rector": "^1.2",
         "symfony/browser-kit": "^6.2",
         "symfony/css-selector": "^6.2",
         "symfony/debug-bundle": "^6.2",
@@ -67,7 +69,8 @@
             "composer/package-versions-deprecated": true,
             "composer/installers": true,
             "symfony/flex": true,
-            "symfony/runtime": true
+            "symfony/runtime": true,
+            "php-http/discovery": true
         }
     },
     "autoload": {
