@@ -8,34 +8,34 @@ use Doctrine\ORM\Mapping as ORM;
 use JMS\Serializer\Annotation as JMS;
 
 #[ORM\Table(name: 'city')]
-#[ORM\Entity(repositoryClass: 'App\Repository\CityRepository')]
+#[ORM\Entity(repositoryClass: \App\Repository\CityRepository::class)]
 #[JMS\ExclusionPolicy('ALL')]
 class City implements \Stringable
 {
     #[ORM\Id]
     #[ORM\Column(type: 'integer')]
     #[ORM\GeneratedValue(strategy: 'AUTO')]
-    protected $id;
+    protected ?int $id = null;
 
     #[ORM\Column(type: 'datetime', nullable: false)]
     #[JMS\Expose]
     #[JMS\Type("DateTime<'U'>")]
-    protected $createdAt;
+    protected ?\DateTime $createdAt = null;
 
     #[ORM\Column(type: 'string', nullable: false)]
     #[JMS\Expose]
-    protected $name;
+    protected ?string $name = null;
 
     #[ORM\Column(type: 'string', nullable: false)]
     #[JMS\Expose]
-    protected $slug;
+    protected ?string $slug = null;
 
     #[ORM\Column(type: 'string', nullable: true)]
     #[JMS\Expose]
-    protected $description;
+    protected ?string $description = null;
 
     #[ORM\OneToMany(targetEntity: 'Station', mappedBy: 'city')]
-    protected $stations;
+    protected Collection $stations;
 
     public function __construct()
     {
@@ -122,6 +122,7 @@ class City implements \Stringable
         return $this;
     }
 
+    #[\Override]
     public function __toString(): string
     {
         return $this->name ?: '';
