@@ -5,6 +5,8 @@ namespace App\Air\AirQuality\LevelColorCollection;
 use App\Air\AirQuality\LevelColors\LevelColorsInterface;
 use App\Air\AirQuality\LevelColors\StandardLevelColors;
 use App\Util\ClassUtil;
+use Symfony\Component\String\ByteString;
+use function Symfony\Component\String\b;
 
 class LevelColorCollection implements LevelColorCollectionInterface
 {
@@ -12,19 +14,18 @@ class LevelColorCollection implements LevelColorCollectionInterface
 
     public function addLevelColors(LevelColorsInterface $levelColors): LevelColorCollectionInterface
     {
-        $lowercaseClassName = ClassUtil::getLowercaseShortname($levelColors);
-        $identifier = str_replace('levelcolors', '', $lowercaseClassName);
-
-        $this->levelColorsList[$identifier] = $levelColors;
+        $this->levelColorsList[$levelColors->getIdentifier()] = $levelColors;
 
         return $this;
     }
 
+    #[\Override]
     public function getLevelColorsList(): array
     {
         return $this->levelColorsList;
     }
 
+    #[\Override]
     public function getLevelColorsForMeasurement(string $measurementIdentifier): LevelColorsInterface
     {
         if (!array_key_exists($measurementIdentifier, $this->levelColorsList)) {
