@@ -5,7 +5,6 @@ namespace App\Air\ViewModelFactory;
 use App\Air\AirQuality\Calculator\AirQualityCalculatorInterface;
 use App\Air\Measurement\CO;
 use App\Air\Measurement\CO2;
-use App\Air\Measurement\CoronaIncidence;
 use App\Air\Measurement\MeasurementInterface;
 use App\Air\Measurement\NO2;
 use App\Air\Measurement\O3;
@@ -14,6 +13,7 @@ use App\Air\Measurement\PM25;
 use App\Air\Measurement\SO2;
 use App\Air\Measurement\Temperature;
 use App\Air\Measurement\UVIndex;
+use App\Air\Measurement\UVIndexMax;
 use Caldera\GeoBasic\Coord\CoordInterface;
 use Caldera\GeoBasic\Coordinate\CoordinateInterface;
 
@@ -27,6 +27,7 @@ abstract class AbstractMeasurementViewModelFactory implements MeasurementViewMod
     {
     }
 
+    #[\Override]
     public function setPollutantList(array $pollutantList): MeasurementViewModelFactoryInterface
     {
         $this->pollutantList = $pollutantList;
@@ -34,6 +35,7 @@ abstract class AbstractMeasurementViewModelFactory implements MeasurementViewMod
         return $this;
     }
 
+    #[\Override]
     public function getPollutantList(): array
     {
         return $this->pollutantList;
@@ -51,10 +53,11 @@ abstract class AbstractMeasurementViewModelFactory implements MeasurementViewMod
             case 7: return new CO2();
             case 8: return new UVIndex();
             case 9: return new Temperature();
-            case 10: return new CoronaIncidence();
+            case 11: return new UVIndexMax();
         }
     }
 
+    #[\Override]
     public function setCoord(CoordinateInterface $coord): MeasurementViewModelFactoryInterface
     {
         $this->coord = $coord;
@@ -62,5 +65,6 @@ abstract class AbstractMeasurementViewModelFactory implements MeasurementViewMod
         return $this;
     }
 
+    #[\Override]
     abstract public function decorate(): MeasurementViewModelFactoryInterface;
 }
