@@ -8,24 +8,31 @@ use App\Entity\Station;
 use App\Pollution\ValueDataConverter\ValueDataConverter;
 use App\Provider\OpenWeatherMapProvider\SourceFetcher\Parser\JsonParserInterface as OwmJsonParserInterface;
 use App\Provider\OpenWeatherMapProvider\SourceFetcher\SourceFetcher as OwmSourceFetcher;
-use App\Provider\CoronaProvider\SourceFetcher\Parser\JsonParserInterface as CoronaParserInterface;
-use App\Provider\CoronaProvider\SourceFetcher\SourceFetcher as CoronaSourceFetcher;
+use App\Provider\OpenUvIoProvider\SourceFetcher\Parser\JsonParserInterface as OpenUvIoJsonParserInterface;
+use App\Provider\OpenUvIoProvider\SourceFetcher\SourceFetcher as OpenUvIoSourceFetcher;
 use Caldera\GeoBasic\Coord\CoordInterface;
 
 class AdhocDataRetriever implements DataRetrieverInterface
 {
-    public function __construct(protected OwmSourceFetcher $owmSourceFetcher, protected OwmJsonParserInterface $owmJsonParser, protected CoronaParserInterface $coronaJsonParser, protected CoronaSourceFetcher $coronaSourceFetcher)
+    public function __construct(
+        protected OwmSourceFetcher $owmSourceFetcher,
+        protected OwmJsonParserInterface $owmJsonParser,
+        protected OpenUvIoJsonParserInterface $openUvIoJsonParser,
+        protected OpenUvIoSourceFetcher $openUvIoSourceFetcher
+    )
     {
+
     }
 
+    #[\Override]
     public function retrieveDataForCoord(CoordInterface $coord, int $pollutantId = null, \DateTime $fromDateTime = null, \DateInterval $dateInterval = null, float $maxDistance = 20.0, int $maxResults = 250): array
     {
         if ($coord instanceof Station) {
             return [];
         }
 
-        if (MeasurementInterface::MEASUREMENT_UVINDEX === $pollutantId) {
-            $data = $this->retrieveUVIndexForCoord($coord);
+        if (MeasurementInterface::MEASUREMENT_UVINDEXMAX === $pollutantId) {
+            $data = $this->retrieveUVIndexMaxForCoord($coord);
 
             if (!$data) {
                 return [];
@@ -44,24 +51,14 @@ class AdhocDataRetriever implements DataRetrieverInterface
             return [$data];
         }
 
-        if (MeasurementInterface::MEASUREMENT_CORONAINCIDENCE === $pollutantId) {
-            $data = $this->retrieveCoronaIncidenceForCoord($coord);
-
-            if (!$data) {
-                return [];
-            }
-
-            return [$data];
-        }
-
         return [];
     }
 
-    protected function retrieveUVIndexForCoord(CoordInterface $coord): ?Data
+    protected function retrieveUVIndexMaxForCoord(CoordInterface $coord): ?Data
     {
-        $jsonData = $this->owmSourceFetcher->queryUVIndex($coord);
+        $jsonData = $this->openUvIoSourceFetcher->queryUVMaxIndex($coord);
 
-        $value = $this->owmJsonParser->parseUVIndex($jsonData);
+        $value = $this->openUvIoJsonParser->parseUVIndexMax($jsonData);
 
         $station = new Station($coord->getLatitude(), $coord->getLongitude());
         return ValueDataConverter::convert($value, $station);
@@ -72,16 +69,6 @@ class AdhocDataRetriever implements DataRetrieverInterface
         $jsonData = $this->owmSourceFetcher->queryTemperature($coord);
 
         $value = $this->owmJsonParser->parseTemperature($jsonData);
-
-        $station = new Station($coord->getLatitude(), $coord->getLongitude());
-        return ValueDataConverter::convert($value, $station);
-    }
-
-    protected function retrieveCoronaIncidenceForCoord(CoordInterface $coord): ?Data
-    {
-        $jsonData = $this->coronaSourceFetcher->queryCoronaIncidence($coord);
-
-        $value = $this->coronaJsonParser->parseCoronaIncidence($jsonData);
 
         $station = new Station($coord->getLatitude(), $coord->getLongitude());
         return ValueDataConverter::convert($value, $station);
