@@ -8,15 +8,18 @@ class ChainedDataRetriever implements DataRetrieverInterface
 {
     protected array $chain = [];
 
-    public function __construct(CachedElasticDataRetriever $elasticDataRetriever, Co2CachedDataRetriever $co2CachedDataRetriever, AdhocDataRetriever $adhocDataRetriever)
+    public function __construct(
+        PostgisDataRetriever $postgisDataRetriever,
+        AdhocDataRetriever $adhocDataRetriever
+    )
     {
         $this->chain = [
-            $elasticDataRetriever,
-            $co2CachedDataRetriever,
+            $postgisDataRetriever,
             $adhocDataRetriever,
         ];
     }
 
+    #[\Override]
     public function retrieveDataForCoord(CoordInterface $coord, int $pollutantId = null, \DateTime $fromDateTime = null, \DateInterval $dateInterval = null, float $maxDistance = 20.0, int $maxResults = 250): array
     {
         $dataList = [];
