@@ -3,7 +3,7 @@
 namespace App\Util\EntityMerger;
 
 use Doctrine\Common\Annotations\Reader;
-use JMS\Serializer\Annotation\Expose;
+use Symfony\Component\Serializer\Annotation\Ignore;
 
 class EntityMerger implements EntityMergerInterface
 {
@@ -11,6 +11,7 @@ class EntityMerger implements EntityMergerInterface
     {
     }
 
+    #[\Override]
     public function merge(object $source, object $destination): object
     {
         $reflectionClass = new \ReflectionClass($source);
@@ -45,12 +46,12 @@ class EntityMerger implements EntityMergerInterface
         $propertyAnnotations = $this->annotationReader->getPropertyAnnotations($reflectionProperty);
 
         foreach ($propertyAnnotations as $propertyAnnotation) {
-            if ($propertyAnnotation instanceof Expose) {
-                return true;
+            if ($propertyAnnotation instanceof Ignore) {
+                return false;
             }
         }
 
-        return false;
+        return true;
     }
 
     protected function generateSetMethodName(\ReflectionProperty $reflectionProperty): string
