@@ -6,7 +6,7 @@ class StringUtil
 {
     public static function camelCaseToUnderscore(string $input, bool $avoidDoubleUnderscore = true): string
     {
-        $output = strtolower(preg_replace('/(?<!^)[A-Z]/', '_$0', $input));
+        $output = strtolower((string) preg_replace('/(?<!^)[A-Z]/', '_$0', $input));
 
         if ($avoidDoubleUnderscore) {
             $output = str_replace('__', '_', $output);
